==> Master_Thesis/3. Code/1a. CSPP holdings - download portfolio data.ipynb <==
## CSPP holdings: Portfolio Data

<p>The ECB publishes list of securities held by it under CSPP on weekly basis. For this research, we are downloading the data for corporate bonds holdings as of Feburary 28, 2022.</p>

<p> <em>Chrome version requirement: 100.0.4896.127 </em></p>
<p> <em>Empty the folder "ECB Data" (2. Data --> 1. ECB data) before running the code. </em></p>

### Import libraries

from selenium.webdriver.chrome.options import Options
from selenium import webdriver
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning) 
from selenium.webdriver.common.action_chains import ActionChains
import time
from selenium.webdriver.common.keys import Keys
import datetime
from selenium.common.exceptions import NoSuchElementException
import os
from functions_file_1 import download_data

### Load the website

directory = os.getcwd()
# specify the fields to load the website
url = "https://www.ecb.europa.eu/mopo/implement/app/html/index.en.html#cspp"
driver_path = directory + '/chromedriver'
download_path = directory[:-7] + "2. Data/1. ECB Data/Downloaded_Data - Web Scrapping"

options = Options()
preferences = {"profile.default_content_settings.popups": 0,
                "download.default_directory": download_path ,
                "directory_upgrade": True}
options.add_experimental_option('prefs', preferences)
driver = webdriver.Chrome(options=options, executable_path=driver_path)
driver.maximize_window()
driver.get(url)
time.sleep(2)
driver.find_element(By.XPATH,"//*[@id='cookieConsent']/div[1]/div/a[2]").click()
time.sleep(2)

### Selection of elements

x = driver.find_element(By.XPATH,"//*[@id='main-wrapper']/main/div[7]/div[1]/div")
build = ActionChains(driver)
build.move_to_element(x).perform()
time.sleep(2)
x2 = driver.find_element(By.XPATH,"//*[@id='main-wrapper']/main/div[7]/div[1]/div")
x2.click()
time.sleep(2)
content_box = driver.find_element(By.XPATH,"//*[@id='main-wrapper']/main/div[7]/div[2]")
WebDriverWait(driver, 2).until(EC.element_to_be_clickable,content_box)
driver.execute_script("arguments[0].click();arguments[0].style.visibility = 'visible';arguments[0].style.display = 'block';",content_box)


### Download data

<p>Data from inception till Feburary 28, 2022</p>

current_date = datetime.datetime.today()
end_date = datetime.datetime(2022,2,25) # fixed data till Feburary
start_date = datetime.datetime(2017,6,23)
n = (current_date. year -end_date.year) * 12 + (current_date. month - end_date. month)
if n > 0:
    for i in range(n):
        prev_arrow = driver.find_element(By.CLASS_NAME,'ui-datepicker-prev.ui-corner-all')
        WebDriverWait(driver, 10).until(EC.element_to_be_clickable,prev_arrow)
        driver.execute_script("arguments[0].click();arguments[0].style.visibility = 'visible';arguments[0].style.display = 'block';",prev_arrow)

# number_of_times_prev_arrow_to_be_clicked
num_months = (end_date. year - start_date.year) * 12 + (end_date. month - start_date. month)
    

# for current month
for x in range(6):
    try:
        download_data(driver,x)
    except NoSuchElementException:
        pass

for i in range(num_months):
    prev_arrow = driver.find_element(By.CLASS_NAME,'ui-datepicker-prev.ui-corner-all')
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable,prev_arrow)
    driver.execute_script("arguments[0].click();arguments[0].style.visibility = 'visible';arguments[0].style.display = 'block';",prev_arrow)
    for x in range(6):
        try:
            year = driver.find_element(By.CLASS_NAME,'ui-datepicker-year')
            month = driver.find_element(By.CLASS_NAME, 'ui-datepicker-month')
            if month.get_attribute('value')=='11' and x == 4 and year.get_attribute('value') == '2021':
                        print("No Data found for December 24,2021") 
            else: 
                download_data(driver, x)
            
        except NoSuchElementException:
            pass
