==> Master_Thesis/3. Code/1b. CSPP holdings - Combined portfolio data (2017-2022).ipynb <==
## Final ECB portfolio data

<p>The list of securities held by ECB under CSPP was downloaded in file "1a. CSPP holdings - download portfolio data" for every week. This file combines them into one, to create a portfolio of CSPP holdings.  </p>

### Import libraries

from functions_file_1 import read_file_ecb,names_of_file
import os
import pandas as pd

directory = os.getcwd()
# path for ECB portfolio data
url = directory[:-7] + "2. Data/1. ECB Data/Downloaded_Data - Web Scrapping/"

### CSPP holdings database

# ECB DATA 
data = []

files_ecb = names_of_file(url)

for f in files_ecb:
    data2 = read_file_ecb(f[1],url)
    data.append(data2)
    
data = pd.concat(data)

# Cleaning data 
ecb_data = data.drop(columns = [5,6])
columns = ['NCB','ISIN_CODE','ISSUER_NAME','MATURITY_DATE','COUPON_RATE','PUBLISHED_DATE','YEAR','MONTH']
ecb_data.columns = columns
ecb_data = ecb_data.dropna(axis = 0)
ecb_data.head()

### Download database

ecb_data.to_excel(directory[:-7] + '2. Data/1. ECB Data/ECB_Portfolio_2017_2022.xlsx',index = None)
