==> Master_Thesis/3. Code/2a. Company Information - Data collection.ipynb <==
## Company Information Data: Revenue and Industry Classification

<p>The revenue and industry classification information is downloaded from S$&$P Capital IQ Pro Platform. However, KU Leuven did not have access to the API of S$&$P, therefore the data had to be downloaded manually. Only 65,000 rows can be extracted from the online Screener at one time, so we had to repeat the step multiple times.</p>

### Libraries to import

from functions_file_2 import bonds_file, read_file_bonds
import os
import pandas as pd

directory = os.getcwd()
# # path for bonds data
path = directory[:-7] + "2. Data/2. Company Information Data/1. Revenue_information - S & P/"
path_2 = directory[:-7] + "2. Data/2. Company Information Data/2. Industry classification information - S & P/"

### Revenue information file

# REVENUE DATA (2017)- S&P
files_rev = bonds_file(path)

data_2 = []
for f in files_rev:
    data2 = read_file_bonds(f[1],path)
    data_2.append(data2)

data_2 = pd.concat(data_2)

columns = ['ISSUER_NAME','INSTRUMENT_ID','ISIN','AMT_OUTSTANDING','ISSUE_PRICE','ISSUE_AMOUNT','TOTAL_REVENUE','TOTAL_DEBT','TOTAL_PREFERRED_EQUITY','TOTAL_COMMON_EQUITY','TOTAL_MINORITY_INTEREST',
           'MARKET_CAPITALIZATION','MATURITY_DATE']
revenue_data = data_2.copy()
revenue_data.columns = columns
revenue_data.head()

### Industry Classification file

files_industry = bonds_file(path_2)

data_2 = []
for f in files_industry:
    data2 = read_file_bonds(f[1],path_2)
    data_2.append(data2)

data_2 = pd.concat(data_2)
columns = ['ISSUER_NAME','INSTRUMENT_ID','ISIN','PARENT_COMPANY_NAME','ULTIMATE_PARENT_COMPANY','INDUSTRY_CLASSIFICATION','SIC_CODE','MATURITY_DATE']

industry_data = data_2.copy()
industry_data.columns = columns
industry_data.head()


### Download file

revenue_data.to_excel(directory[:-7] + '2. Data/2. Company Information Data/Combined_Revenue_S&P.xlsx',index = None)


industry_data.to_excel(directory[:-7] + '2. Data/2. Company Information Data/Combined_Industry_S&P.xlsx',index = None)
