==> Master_Thesis/3. Code/2b. Company Information - with CSPP holdings.ipynb <==
## Company Information data combined with ECB portfolio data

<p>The revenue, industry classification and country of risk information data was merged with the portfolio data to gather all information in one file. The revenue was also converted into million dollars format as it would be required to calculate carbon footprint.</p>

### Libraries to import

import datetime as dt
import pandas as pd
import os
from functions_file_2 import clean_file,clean_file_rev,clean_file_industry,merge_revenue,merge_country,clean_country


directory = os.getcwd()
url = directory[:-7] + "2. Data/"

### Import Dataset

ecb_portfolio = pd.read_excel(url + "1. ECB Data/ECB_Portfolio_2017_2022.xlsx")

# revenue information
bonds_snp = pd.read_excel(url + "2. Company Information Data/Combined_Revenue_S&P.xlsx" )
mapping_revenue = pd.read_excel(url + "2. Company Information Data/3. Mapping/Mapping.xlsx",sheet_name = "Bonds_Revenue_Mapping",skiprows=4)


# industry classification information
sector_info = pd.read_excel(url + "2. Company Information Data/3. Mapping/Mapping.xlsx",sheet_name = "Bonds_SIC_Mapping",skiprows=4)


# country of risk
country_risk_data = pd.read_excel(url + "2. Company Information Data/Company's Country of Risk Info.xlsx")
mapping_country = pd.read_excel(url + "2. Company Information Data/3. Mapping/Mapping.xlsx",sheet_name = "Country_Mapping",skiprows=4)


### Portfolio with revenue

# clean the bonds data
bonds_snp = clean_file(bonds_snp)

# merge ecb portfolio with revenue data from S&P and mapped file
df = merge_revenue(ecb_portfolio,bonds_snp,mapping_revenue)

# Portfolio with total revenue
df_revenue = clean_file_rev(df)

#### Converting revenue to million format

df_revenue['TOTAL__REVENUE'] = (df_revenue['TOTAL__REVENUE'].apply(lambda x: x/1000)).round(2)
df_revenue.head()

### Portfolio with Industry Classification

sector_info = sector_info[['Unique_ISIN_Code','SIC_CODE','NACE_CODE','ECONOMIC_SECTOR']]


# merge with industry classification information
subset = ['NCB', 'ISIN_CODE', 'ISSUER_NAME', 'MATURITY_DATE', 'COUPON_RATE','PUBLISHED_DATE', 'YEAR', 'MONTH', 'TOTAL__REVENUE']
merge_sector = df_revenue.merge(sector_info,left_on = ['ISIN_CODE'],right_on= ['Unique_ISIN_Code'], how = "left").drop_duplicates(subset = subset)


# data includes additional information about SIC Code, NACE Code, Economic Sector
df_company_info = clean_file_industry(merge_sector)

df_company_info.head()

#### Filtering out data with no revenue information

<p> Revenue information for year 2017 was not available for 155 individual companies, a total of 59461 rows (18% of the total data). These rows were removed from the database. A company named 'Delta Lloyd NV' had negative revenue was also removed.</p>

# no revenue information available
print(df_company_info[df_company_info['TOTAL__REVENUE'] <= 0]['ISSUER_NAME'].nunique())

# removing rows no revenue information
final_data_co_info = df_company_info[df_company_info['TOTAL__REVENUE'] > 0]

### Portfolio with country of risk

data = merge_country(final_data_co_info,country_risk_data,mapping_country)

df = clean_country(data)

### Classifying dates into quarters

data = df.copy()
data['QUARTER'] = df['PUBLISHED_DATE'].dt.to_period('Q')
data.head()

### Download file

data.to_excel(directory[:-7] + "2. Data/ECB Portfolio - Company Information.xlsx",index=None)
