==> Master_Thesis/3. Code/2c. Cleaned Company Information with portfolio data.ipynb <==
## Company Information with portfolio data (cleaning)

<p>The list of securities published by the ECB had few issues. The same ISIN code had different issuer names that was checked and corrected in our data.  </p>
<p>The company information data was collected from different sources, so there were few discripancies in the data. For instance, same company names had different industry classification and different revenue. This information was checked manually from Orbis database and corrected in our data. </p>

### Libraries to import

import pandas as pd
import numpy as np
import os
from functions_file_2 import clean_issuer_info

directory = os.getcwd()
path = directory[:-7] + "2. Data/"

### Import datasets

file = pd.read_excel(path + "ECB Portfolio - Company Information.xlsx")

# correct informaation datasets
correct_issuer_name = pd.read_excel(path + "2. Company Information Data/3. Mapping/Mapping.xlsx",sheet_name = "Correct Issuer Name",skiprows=4)
correct_nace = pd.read_excel(path + "2. Company Information Data/3. Mapping/Mapping.xlsx",sheet_name = "Correct NACE Code",skiprows=4)
correct_revenue = pd.read_excel(path + "2. Company Information Data/3. Mapping/Mapping.xlsx",sheet_name = "Correct Revenue ",skiprows=4)


### Map datasets with the correct information

#### Issuer Name

# merge with correct issuer name file
df = file.merge(correct_issuer_name,left_on ="ISSUER_NAME",right_on = "Actual name", how = "left").drop(['Actual name'],axis=1)


df['ISSUER_NAME'] = np.where(df['Correct mapping'].isna(),df['ISSUER_NAME'],df['Correct mapping'])
dataset = df.drop(['Correct mapping'],axis = 1)

df = clean_issuer_info(dataset)

#### Industry classification

df2 = df.merge(correct_nace,on = "ISSUER_NAME",how = "left")

df2['SIC_CODE'] = np.where(df2['SIC '].isna(),df2['SIC_CODE'],df2['SIC '])
df2['NACE_CODE'] = np.where(df2['NACE'].isna(),df2['NACE_CODE'],df2['NACE'])
df2['ECONOMIC_SECTOR'] = np.where(df2['ECONOMIC SECTOR'].isna(),df2['ECONOMIC_SECTOR'],df2['ECONOMIC SECTOR'])

columns_to_drop = ['SIC ','NACE','ECONOMIC SECTOR']
df2 = df2.drop(columns_to_drop,axis = 1)

#### Revenue

df3 = df2.merge(correct_revenue,on='ISSUER_NAME',how ="left")
df3['TOTAL__REVENUE'] = np.where(df3['REVENUE'].isna(),df3['TOTAL__REVENUE'],df3['REVENUE'])
final_data = df3.drop('REVENUE',axis = 1)

final_data.head()

### Download file

final_data.to_excel(path + "ECB Portfolio - Company Information.xlsx",index=None)
