==> Master_Thesis/3. Code/3a. Emissions' Data - CDP.ipynb <==
## CDP Emissions Data, 2017

<p>The Scope 1 and Scope 2 emissions data is collected from CDP for year 2017. The gross global year emissions were used because they correspond to the recent reported year. Location-based Scope 2 emissions were used because they are lots of issues with market-based method. </p>

### Libraries to import

import pandas as pd
import os
from functions_file_3 import clean_scope1, clean_scope2, finalizedata

directory = os.getcwd()
url = directory[:-7] + "2. Data/3. Emissions Data - CDP/GHG_Emissions_Data - CDP/"

### Import Datasets

# Company information in the CDP data
company_info = pd.read_excel(url + "Investor CDP 2017_Public Data (with sector Qs).xlsx", sheet_name = "Summary Data")

# Gross Global Scope 1 and Scope 2 emissions
emissions_scope1 = pd.read_excel(url + "Investor CDP 2017_Public Data (with sector Qs).xlsx", sheet_name = "CC8. Emissions Data",skiprows = 1,header = None)
emissions_scope2 = pd.read_excel(url + "Investor CDP 2017_Public Data (with sector Qs).xlsx", sheet_name = "CC8.3a",skiprows = 2, header = None)


### Preprocessing

info_data = company_info[['account_id','account_name','isin']]
columns = ['ACCOUNT_ID','ACCOUNT_NAME','ISIN']
info_data.columns = columns 

scope1_data = clean_scope1(emissions_scope1)
scope2_data = clean_scope2(emissions_scope2)

### Emissions Data for CDP companies

emissions_data = finalizedata(info_data, scope1_data,scope2_data)
emissions_data.head()

### Download file

emissions_data.to_excel(directory[:-7] + "2. Data/3. Emissions Data - CDP/Emissions_Data.xlsx",index = None)
