==> Master_Thesis/3. Code/3b. Emissions - Intensity and Master File.ipynb <==
## Emission Intensity & Final Database

<p>The ECB portfolio is mapped with the CDP emissions data. However, out of 294 companies, 112 companies' emissions were not available in the data. To solve this, emission intensity was calculated based on the sector in proportion to the revenue of the companies. </p>

### Libraries to import

import pandas as pd
import os
import warnings
warnings.filterwarnings("ignore")
from functions_file_3 import match_data,match_emissions,match_emissions_cdp,intensity,emissions_portfolio
pd.options.display.float_format = '{:.2f}'.format

directory = os.getcwd()
url = directory[:-7] + '2. Data/'

### Import datasets

# import ECB portfolio with all company information
portfolio = pd.read_excel(url + 'ECB Portfolio - Company Information.xlsx')

# emissions data
emissions_data = pd.read_excel(url + '3. Emissions Data - CDP/Emissions_Data.xlsx')

# import mappings to calculate emission intensity
dictionary_for_cdp = pd.read_excel(url + '3. Emissions Data - CDP/Emissions_Mapping.xlsx',sheet_name = 'Dictionary_for_CDP',skiprows=4)
cdp_companies = pd.read_excel(url + '3. Emissions Data - CDP/Emissions_Mapping.xlsx',sheet_name = 'CDP_Companies',skiprows = 4,usecols="A:G")


### Companies matched with CDP emissions data

matched_data, unmatched_data = match_data(portfolio,dictionary_for_cdp)

matched_emissions = match_emissions(matched_data,emissions_data)

matched_emissions.head()

### Emission Intensity based on NACE sector code

# merge emissions data with companies in CDP data
cdp_emissions = match_emissions_cdp(cdp_companies,emissions_data)

# calculate emissions intensity based on NACE code of sectors
emission_intensity = intensity(cdp_emissions)

### Companies unmatched with CDP emissions data

# unmatched data with emission intensity
data = unmatched_data.merge(emission_intensity, on = "NACE_CODE", how = "left")

### ECB Portfolio with emissions data

final_data = emissions_portfolio(portfolio, matched_emissions,data)
final_data.head()

### Download the file

final_data.to_excel(url + 'Master_File.xlsx',index = None)
