==> Master_Thesis/3. Code/4d. Yearly WACI.ipynb <==
## Yearly Weighted Average Carbon Intensity (WACI)

<p>The WACI has two main components to it: emissions expressed as number of tonnes of CO2 equivalent and firms' million dollar revenue that is used to normalize the metric. The WACI is calculated yearly for every sector and country of risk by taking the average of monthly portfolio values and sectoral and country percentage breakdown. It is not calculate for individual firms as ECB does not publish the volume of each bonds' issue it purchases.  </p>

### Libraries to import

import pandas as pd
import os
from functions_file_3 import risk_country, yearly_waci_calculation_file,clean_master_file_yearly,yearly_mean_weighted_carbon_intensity, yearly_median_weighted_carbon_intensity, check_difference_waci,create_data,yearly_waci_sector_calculation,contribution_calculate
import warnings
warnings.filterwarnings("ignore")

directory = os.getcwd()
url = directory[:-7] + '2. Data/'

### Import datasets

master_file = pd.read_excel(url + 'Master_file.xlsx')

portfolio_value = pd.read_excel(url + '1. ECB Data/Portfolio_Breakdown_Information.xlsx',
                               sheet_name = 'Yearly_Portfolio_Value')
sector_breakdown = pd.read_excel(url + '1. ECB Data/Portfolio_Breakdown_Information.xlsx',
                               sheet_name = 'Yearly_Sector_Breakdown')
country_breakdown = pd.read_excel(url + '1. ECB Data/Portfolio_Breakdown_Information.xlsx',
                               sheet_name = 'Yearly_Country_Breakdown')

### Preprocessing

df = clean_master_file_yearly(master_file)

final_data = yearly_waci_calculation_file(df,portfolio_value,sector_breakdown,country_breakdown)
final_data = final_data.dropna()
final_data.head()

### Yearly Sector WACI

df2 = yearly_waci_sector_calculation(df,portfolio_value,sector_breakdown)

yearly_sector_waci = contribution_calculate(df2)
yearly_sector_waci.head()

### Yearly Mean WACI and Median WACI dataset

final_data['WEIGHTS'] = (final_data['SECTOR_CSPP_HOLDINGS']* final_data['COUNTRY_CSPP_HOLDINGS']*final_data['TOTAL_HOLDINGS(USD)'])/final_data['TOTAL_HOLDINGS(USD)']
final_data['SUM_WEIGHTS'] = final_data.groupby('YEAR')['WEIGHTS'].transform('sum')
final_data['MEAN_WACI'] = final_data['WEIGHTS']*final_data['MEAN_WEIGHTED_INTENSITY']/final_data['SUM_WEIGHTS']
final_data['MEDIAN_WACI'] = final_data['WEIGHTS']*final_data['MED_WEIGHTED_INTENSITY']/final_data['SUM_WEIGHTS']


columns_to_keep = ['YEAR','ECONOMIC_SECTOR','RISK_COUNTRY','MEAN_WACI','MEDIAN_WACI']
data = final_data[columns_to_keep]
data.head()

### Determine companies driving mean WACI

keys = [2017,2018,2019,2020,2021,2022]

df=[]
for k in keys:
    sector, condition = check_difference_waci(final_data,k)
    df2 = create_data(master_file,condition)
    df.append(df2)
    
new_data = pd.concat(df)

columns = ['Issuer Name','Risk Country','Economic Sector','Year','Number of bonds','Avg. Emission Intensity']
new_data.columns = columns

### Select a year

year = 2018

yearly_mean_weighted_carbon_intensity(year,final_data)

yearly_median_weighted_carbon_intensity(year,final_data)

### Download file

path = directory[:-7] + '2. Data/4. Dashboard Data/'
data.to_csv(path + 'WACI_data.csv',index = None)
final_data.to_csv(path + 'Final_data.csv',index=None)
new_data.to_csv(path + 'Data.csv',index = None)
yearly_sector_waci.to_csv(path + 'Yearly_sector_WACI')
