==> Master_Thesis/3. Code/5. Datasets for Dashboard.ipynb <==
## Datasets for Dashboard

<p>Datasets required for creating visualisations in the dashboard. </p>

### Libraries to import

import pandas as pd
import os
import datetime as dt
import warnings
warnings.filterwarnings("ignore")
from functions_file_3 import risk_country

directory = os.getcwd()
url = directory[:-7] + '2. Data/1. ECB Data/Portfolio Value Information/'
url_2 = directory[:-7] + '2. Data/1. ECB Data/'
path = directory[:-7] + '2. Data/4. Dashboard Data/'

### Functions to use

def clean_value(filename):
    # selecting rows for which we have downloaded ECB bonds data
    df = filename.iloc[12:,:]
    # calculating portfolio value for every quarter starting from June 2017 to Feb 2022
    df['DATE'] = df['End of Month'].map(lambda x: dt.datetime.strptime(x, '%b-%y'))
    df['DATE'] = df['End of Month'].map(lambda x: dt.datetime.strptime(x, '%b-%y'))
    df['Month'] = pd.to_datetime(df['DATE']).dt.strftime('%b')
    df['Year'] = pd.DatetimeIndex(df['DATE']).year
    df['Total holdings'] = df['Total holdings'].str.replace(",","")
    df['Total holdings'] = df['Total holdings'].apply(pd.to_numeric)
    columns_to_keep = ['Year','Month','Total holdings']
    data = df[columns_to_keep]
    return data

### Monthly CSPP holdings (June 2017 - Feb 2022)

portfolio_value = pd.read_csv(url + 'CSPP_breakdown_history.csv',skiprows=1,skipfooter=2, engine = 'python')

monthly_value = clean_value(portfolio_value)
monthly_value['Month_Number']=[dt.datetime.strptime(x, "%b") for x in monthly_value['Month']]
monthly_value['Month_Number'] = [x.month for x in monthly_value['Month_Number']]
monthly_value.head()

monthly_value.to_csv(path + 'Monthly_Investments.csv')

### Sector wise 

yearly_sector_value = pd.read_excel(url_2 + 'Portfolio_Breakdown_Information.xlsx',sheet_name='Yearly_Sector_Breakdown')
yearly_portfolio = pd.read_excel(url_2 + 'Portfolio_Breakdown_Information.xlsx',sheet_name='Yearly_Portfolio_Value')


combined_data = yearly_sector_value.merge(yearly_portfolio,on = "YEAR", how ="left").drop_duplicates()
combined_data['SECTOR_HOLDINGS'] = (combined_data['SECTOR_CSPP_HOLDINGS']*combined_data['TOTAL_HOLDINGS(USD)']).round(0)
columns_to_keep = ['YEAR','ECONOMIC_SECTOR','SECTOR_HOLDINGS']
sector_data = combined_data[columns_to_keep]
sector_data.head()

### Country wise

yearly_country_value = pd.read_excel(url_2 + 'Portfolio_Breakdown_Information.xlsx',sheet_name='Yearly_Country_Breakdown')


combined_data = yearly_country_value.merge(yearly_portfolio,on = "YEAR", how ="left").drop_duplicates()
combined_data['COUNTRY_HOLDINGS'] = (combined_data['COUNTRY_CSPP_HOLDINGS']*combined_data['TOTAL_HOLDINGS(USD)']).round(0)
columns_to_keep = ['YEAR','RISK_COUNTRY','COUNTRY_HOLDINGS']
country_data = combined_data[columns_to_keep]
country_data.head()

country_data.to_csv(path + 'Country_values.csv',index = None)
sector_data.to_csv(path + 'Sector_values.csv',index = None)

### Carbon Intensity for individual companies

CI_companies = pd.read_csv(path + 'CI_for_companies.csv')
master_file = pd.read_excel(directory[:-7] + '2. Data/Master_File.xlsx')

# count the number of bonds for every year
columns_to_keep = ['ISSUER_NAME','YEAR','ISIN_CODE']
file = master_file[columns_to_keep]
df = file.groupby(by = ['YEAR','ISSUER_NAME']).count()
df = df.reset_index()
columns = ['YEAR','ISSUER_NAME','NUMBER_OF_BONDS']
df.columns = columns

# combine with bonds info with carbon intensity of companies
df2 = df.merge(CI_companies, on = 'ISSUER_NAME',how = 'left').drop_duplicates()
df2 = df2.dropna()


carbon_intensity_companies = df2.sort_values(by= 'CARBON_INTENSITY',ascending = False)
carbon_intensity_companies.head()

carbon_intensity_companies.to_csv(path + 'Carbon_Intensity_Companies.csv',index = None)

### Bonds distribution per sector

# number of bonds in each sector
cols = ['YEAR','ECONOMIC_SECTOR','ISIN_CODE']
file2 = master_file[cols]

df2 = file2.groupby(['YEAR','ECONOMIC_SECTOR']).count()
df2 = df2.reset_index()
columns = ['YEAR','ECONOMIC_SECTOR','NUMBER_OF_BONDS']
df2.columns = columns
df2.head()

df2.to_csv(path + 'Bonds_sectorwise.csv',index=None)

### Bonds distribution per country

file3 = risk_country(master_file)

cols = ['YEAR','RISK_COUNTRY','ISIN_CODE']
file3 = file3[cols]

df3 = file3.groupby(['YEAR','RISK_COUNTRY']).count()
df3 = df3.reset_index()
columns = ['YEAR','RISK_COUNTRY','NUMBER_OF_BONDS']
df3.columns = columns
df3.head()


df3.to_csv(path + 'Bonds_countrywise.csv',index = None)
