==> Master_Thesis/3. Code/6. Datasets for Network Analysis.ipynb <==
## Datasets used for creating a network

<p>Four datasets are created:
      a) Relationship data: ECB buys bonds;
      b) Company nodes data: consists unique names of companies along with industry classification;
      c) Sector nodes data: consists unique sector names;
      d) Country of risk nodes data: consists unique names of country of risk.</p>

### Libraries to import

import pandas as pd
import os

directory = os.getcwd()
path = directory[:-7] + '4. Network Analysis/1. Data/'

### Import dataset

portfolio = pd.read_excel(path + 'ECB Portfolio - Company Information.xlsx')

### a) Relationship data

# functions
def create_relations(data):
    df = data[['ISIN_CODE','QUARTER','ISSUER_NAME']]
    rel = df.groupby(['ISIN_CODE','QUARTER'])['ISSUER_NAME'].count()
    rel = rel.reset_index()
    rel['from'] = 'ECB'
    rel['type'] = 'Buys'
    rel = rel.rename(columns = {'ISSUER_NAME':'number_of_issues'})
    return rel

def clean_data(file):
    columns_to_keep = ['ISIN_CODE','ISSUER_NAME']
    company_name = file[columns_to_keep].drop_duplicates()
    company_name['ISSUER_NAME']=company_name['ISSUER_NAME'].astype(str)
    df = company_name.groupby('ISIN_CODE')['ISSUER_NAME'].apply(';'.join).reset_index()
    df[['ISSUER_NAME_1','ISSUER_NAME_2']]= df['ISSUER_NAME'].str.split(';', expand=True)
    df = df.drop(columns = 'ISSUER_NAME',axis = 1)
    df = df.fillna('-')
    df2 = file[['ISIN_CODE','SECTOR','NACE_CODE','COUNTRY_OF_RISK','COUPON_RATE','MATURITY_DATE']].drop_duplicates()
    isin_info =df.merge(df2,on='ISIN_CODE',how='left').drop_duplicates(subset = df.columns)
    return isin_info
    

relation = create_relations(portfolio)
isin_info_data = clean_data(portfolio)

relationship_data = relation.merge(isin_info_data,on ='ISIN_CODE',how = "left").drop_duplicates()

relationship_data.head()

### b) Company nodes data

columns = ['ISSUER_NAME','NACE_CODE','SECTOR','COUNTRY_OF_RISK']
company = portfolio[columns]
company = company.drop_duplicates(subset = columns)
company.head()

### c) Sector nodes data

columns = ['NACE_CODE','SECTOR']
sector = portfolio[columns]
sector = sector.drop_duplicates(subset = columns)
sector.head()

### d) Country of risk nodes data

country = portfolio[['COUNTRY_OF_RISK']]
country = country.drop_duplicates(subset = ['COUNTRY_OF_RISK'])
country.head()

### Download files

relationship_data.to_csv(path + 'relationship_2.csv',index = False)
company.to_csv(path + 'company_node.csv',index=False)
sector.to_csv(path + 'sector_node.csv',index=False)
country.to_csv(path + 'country_node.csv',index=False)
